requirement.txt
fastapi[all]==0.75.2
redis==4.2.0rc1
cos-python-sdk-v5==1.9.15
PyJWT==2.3.0
python-dotenv==0.19.1
tortoise-orm[aiomysql]==0.19.0
tencentcloud-sdk-python==3.0.618
qcloud-python-sts==3.1.2
qrcode[pil]==7.3.1
passlib==1.7.4
sqlmodel==0.0.6
paho-mqtt==2.1.0
rsa==4.7.2
wechatpayv3==1.2.52
aiomysql==0.2.0
oss2==2.19.0
pyOpenSSL==24.2.1
apscheduler==3.10.4
pillow==10.4.0